Return the same financial year when the start month is January
If the start of the fiscal year is January, the financial year should be the
same whether you're using forward year or not.

Forward year is basically depends on when the fiscal year started (non-forward)
or ended (forward). Since a January start means that the year starts and ends in
the same year, the financial year should be exactly that year.